Properties panel: set cacheBuffer to prevent scrolling issues
Usually, QML ListView will guess the height of the entire list content based on the list items that are currently inside the viewport. In the case of the properties panel, that doesn't work well, because the different sections have very different heights. That can lead to jumping during scrolling. Setting `cacheBuffer` to `contentHeight` prevents this. Note that this is normally completely against the principle of ListView, because it means in practice that all items will always be instantiated instead of only those inside the viewport, but in the case of the properties panel, that's okay, because we're using ListView for its asynchrony, not for its viewport-only efficiency.