Prevent crash using heirarchical popup menus on iOS.
On iOS, a popup with a submenu would be closed when the user selects
the submenu. This leaves the submenu parentless, so it gets deleted while
it is still in use, and we crash.

For iOS, we always leave popup menus selected, this prevents the crash,
but requires the user to hold the tap and slide to the submenu to keep it
open. But the submenu works, and the crash is gone.