Make qt_import_qml_plugins work properly
In Qt 6, it takes only one argument, namely a target, and it scans only the directory of that target, instead of the entire repo like we want. As a hack, we create a fake target at the toplevel directory of the repo, then run `qt_import_qml_plugins` on that target, and then statically link `MuseScoreStudio` to that target.